ENH: update clutter/classify-module/tests
