+++ wradlib/__init__.py
@@ -24,7 +24,6 @@ except Exception:
 from . import adjust  # noqa
 from . import atten  # noqa
 from . import classify  # noqa
-from . import clutter  # noqa
 from . import comp  # noqa
 from . import dp  # noqa
 from . import georef  # noqa
